fix: remove vpermit_tao_limit requirement
Currently size of the network is too small, so every neuron/node of the
network is a validator by default (until we reach 64 neurons/nodes).
Thus, the check for vpermit_tao_limit is set rather small (10 TAO) and
with that, it makes every miner that's not new, a validator (because by
performing same work, it gains enough TAO to be considered validator).